Move ConfirmationDialog into Dialog.features
